requirements.txt
pytubefix
customtkinter
tkinter
pillow